+++ src/pages/sign/signin/index.ts
@@ -1,0 +1,1 @@
+export { default as SignInForm } from "./SignInForm";
