+++ src/pages/sign/signup/index.ts
@@ -1,0 +1,4 @@
+export { default as SignUpForm } from "./SignUpForm";
+export { default as FirstStep } from "./FirstStep";
+export { default as LastStep } from "./LastStep";
+export { default as StepButton } from "./StepButton";
